remodule to improve blockchain sync
1. remodule locate pDeleteBlockIndex
It's time-consuming in SetTip(), what's worse, it's a total waste
to construct a temporary CChain to locate pDeleteBlockIndex, as we
can locate it recursively in a minute.

2. add GetPrunedLocator()
If the peer has the longer chain, it's enough to provide the tip's
hash.

3. add PushGetBlocksWithCondition()
Push `getblocks` request when it's necessary in a fixed and proper
interval.

4. remove unused comments